feature: 0183 TCP server duplex, sentenceEvent
This adds the ability to use the server's NMEA 0183
TCP server for bidirectional communications. The server listens
for 'nmea0183' events, so we will need to be able to
suppress echoing the incoming data back to the client.
Thus we need an option to suppress sending sentences
as 'nmea0183' events.

This also adds also the ability to specify one
sentenceEvent name for emitting sentences from the
0183 parser.